+++ src/app/routes/layouts/RootLayout.tsx
@@ -2,6 +2,8 @@ import { Outlet, ScrollRestoration } from 'react-router-dom';
 import { useFcmForegroundMessage } from '@/shared/lib/firebase/useFcmForegroundMessage';
 import { useUnsavedChangesGuard } from '@/shared/hooks/useUnsavedChangesGuard';
 import { LoginModal } from '@/features/auth/login/ui/LoginModal';
+import { GlobalImageViewer } from '@/shared/ui/elements/modal/image-viewer/ImageViewer';
+import { GlobalAlerts } from '@/shared/ui/elements/modal/alert/Alert';
 
 /**
  * Root Layout
@@ -16,6 +18,8 @@ export function RootLayout() {
       <ScrollRestoration />
       <Outlet />
       <LoginModal />
+      <GlobalImageViewer />
+      <GlobalAlerts />
     </>
   );
 }